+++ samples/TurtlePath.Samples.Basic/Infrastructure/Persistence/CommerceDbContext.cs
@@ -1,5 +1,6 @@
 using Microsoft.EntityFrameworkCore;
 using TurtlePath.EntityFrameworkCore;
+using TurtlePath.EntityFrameworkCore.Conventions;
 using TurtlePath.Samples.Basic.Domain.Entities;
 
 namespace TurtlePath.Samples.Basic.Infrastructure.Persistence;
@@ -8,8 +9,9 @@ public sealed class CommerceDbContext : BaseDbContext
 {
     public CommerceDbContext(
         DbContextOptions<CommerceDbContext> options,
-        TurtlePathDbContextOptions turtlePathOptions)
-        : base(options, turtlePathOptions)
+        TurtlePathDbContextOptions turtlePathOptions,
+        IEnumerable<ITurtlePathModelConvention> modelConventions)
+        : base(options, turtlePathOptions, modelConventions)
     {
     }